# new typecache for better performance