--- a/fastssz.go
+++ b/fastssz.go
@@ -24,109 +24,32 @@ var sszMarshalerType = reflect.TypeOf((*fastsszMarshaler)(nil)).Elem()
 var sszUnmarshalerType = reflect.TypeOf((*fastsszUnmarshaler)(nil)).Elem()
 var sszHashRootType = reflect.TypeOf((*fastsszHashRoot)(nil)).Elem()
 
-// fastsszConvertCompatibility holds information about a type's compatibility with fastssz's static encoding and decoding methods.
-// It is used to determine whether a type can leverage fastssz's efficient, static code paths or if it must be handled dynamically
-// due to the presence of non-default specification values or the lack of necessary interface implementations.
-//
-// Fields:
-//   - isMarshaler: Indicates whether the type implements the fastssz Marshaler interface.
-//   - isUnmarshaler: Indicates whether the type implements the fastssz Unmarshaler interface.
-//   - isHashRoot: Indicates whether the type implements the fastssz HashRoot interface.
-//   - hasDynamicSpecValues: Indicates the presence of dynamically applied specification values that deviate from the default
-//     specifications. A true value here suggests that, despite potentially implementing the required interfaces for static processing,
-//     the type may still need to be handled dynamically due to these spec values affecting its size or structure.
-type fastsszConvertCompatibility struct {
-	isMarshaler         bool
-	isUnmarshaler       bool
-	hasDynamicSpecSizes bool
-}
-
-type fastsszHashCompatibility struct {
-	isHashRoot          bool
-	hasDynamicSpecSizes bool
-	hasDynamicSpecMax   bool
-}
-
 // getFastsszCompatibility evaluates the compatibility of a given type with fastssz, determining whether the type and its nested
-// structures can be efficiently encoded/decoded using fastssz's static code generation approach. This assessment includes checking
-// for the implementation of the Marshaler/Unmarshaler interfaces and the absence of non-default dynamically applied specification
-// values within the type hierarchy. For performance optimization, the results of this compatibility check are cached, allowing
-// `dynssz` to quickly reference these results in future operations without the need to re-evaluate the same types repeatedly.
+// structures can be efficiently encoded/decoded using fastssz's static code generation approach.
 //
 // Parameters:
 // - targetType: The reflect.Type of the value being assessed for fastssz compatibility. This type, along with its nested
 //   or referenced types, is evaluated to ensure it aligns with fastssz's requirements for static encoding and decoding.
 //
 // Returns:
-// - A pointer to a fastsszCompatibility struct, which contains detailed information about the compatibility status, including
-//   whether the type implements required interfaces and lacks dynamically applied non-default spec values.
-// - An error if the compatibility check encounters issues, such as reflection errors or the presence of unsupported type configurations
-//   that would prevent the use of fastssz for encoding or decoding.
+// - A boolean indicating whether the type is compatible with fastssz's static encoding and decoding.
 
-func (d *DynSsz) getFastsszConvertCompatibility(targetType reflect.Type, sizeHints []sszSizeHint) (*fastsszConvertCompatibility, error) {
-	d.fastsszConvertCompatMutex.Lock()
-	defer d.fastsszConvertCompatMutex.Unlock()
-
-	if cachedCompatibility := d.fastsszConvertCompatCache[targetType]; cachedCompatibility != nil {
-		return cachedCompatibility, nil
-	}
-
-	_, hasDynSizes, err := d.getSszSize(targetType, sizeHints)
-	if err != nil {
-		return nil, err
-	}
-
+func (d *DynSsz) getFastsszConvertCompatibility(targetType reflect.Type) bool {
 	targetPtrType := reflect.New(targetType).Type()
-	compatibility := &fastsszConvertCompatibility{
-		isMarshaler:         targetPtrType.Implements(sszMarshalerType),
-		isUnmarshaler:       targetPtrType.Implements(sszUnmarshalerType),
-		hasDynamicSpecSizes: hasDynSizes,
-	}
-	d.fastsszConvertCompatCache[targetType] = compatibility
-	return compatibility, nil
+	return targetPtrType.Implements(sszMarshalerType) && targetPtrType.Implements(sszUnmarshalerType)
 }
 
 // getFastsszHashCompatibility evaluates the compatibility of a given type with fastssz's HashRoot interface, determining whether
-// the type can efficiently compute its hash tree root using fastssz's static code generation approach. This assessment includes
-// checking for the implementation of the HashRoot interface and the absence of non-default dynamically applied specification values
-// within the type hierarchy. For performance optimization, the results of this compatibility check are cached, allowing `dynssz`
-// to quickly reference these results in future operations without the need to re-evaluate the same types repeatedly.
+// the type can efficiently compute its hash tree root using fastssz's static code generation approach.
 //
 // Parameters:
 // - targetType: The reflect.Type of the value being assessed for fastssz hash compatibility. This type, along with its nested
 //   or referenced types, is evaluated to ensure it aligns with fastssz's requirements for static hash tree root computation.
 //
 // Returns:
-// - A pointer to a fastsszHashCompatibility struct, which contains detailed information about the compatibility status, including
-//   whether the type implements the HashRoot interface and lacks dynamically applied non-default spec values that would affect
-//   the hash tree root computation.
-// - An error if the compatibility check encounters issues, such as reflection errors or the presence of unsupported type configurations
-//   that would prevent the use of fastssz for hash tree root computation.
+// - A boolean indicating whether the type is compatible with fastssz's static hash tree root computation.
 
-func (d *DynSsz) getFastsszHashCompatibility(targetType reflect.Type, sizeHints []sszSizeHint, maxSizeHints []sszMaxSizeHint) (*fastsszHashCompatibility, error) {
-	d.fastsszHashCompatMutex.Lock()
-	defer d.fastsszHashCompatMutex.Unlock()
-
-	if cachedCompatibility := d.fastsszHashCompatCache[targetType]; cachedCompatibility != nil {
-		return cachedCompatibility, nil
-	}
-
-	compat, err := d.getFastsszConvertCompatibility(targetType, sizeHints)
-	if err != nil {
-		return nil, err
-	}
-
-	hasDynMaxSize, err := d.checkDynamicMaxSize(targetType, maxSizeHints)
-	if err != nil {
-		return nil, err
-	}
-
+func (d *DynSsz) getFastsszHashCompatibility(targetType reflect.Type) bool {
 	targetPtrType := reflect.New(targetType).Type()
-	compatibility := &fastsszHashCompatibility{
-		isHashRoot:          targetPtrType.Implements(sszHashRootType),
-		hasDynamicSpecSizes: compat.hasDynamicSpecSizes,
-		hasDynamicSpecMax:   hasDynMaxSize,
-	}
-	d.fastsszHashCompatCache[targetType] = compatibility
-	return compatibility, nil
+	return targetPtrType.Implements(sszHashRootType)
 }