… sszmax.go
… sszsize.go
… treeroot.go