old caching system:
```
## mainnet preset / BeaconBlock decode + encode + hash (1000 times)
fastssz only        [   1 ms /    0 ms /    8 ms]        success         Root: 0x13515a87c6e7f37b5b39d6d1b69a0819e93b04bb4e6b3c3d57d14d6b7af3b45d
dynssz only         [  25 ms /   40 ms /   28 ms]        success         Root: 0x13515a87c6e7f37b5b39d6d1b69a0819e93b04bb4e6b3c3d57d14d6b7af3b45d
dynssz + fastssz    [   0 ms /    0 ms /    8 ms]        success         Root: 0x13515a87c6e7f37b5b39d6d1b69a0819e93b04bb4e6b3c3d57d14d6b7af3b45d

## mainnet preset / BeaconState decode + encode + hash (1000 times)
fastssz only        [ 535 ms /  474 ms / 7040 ms]        success         Root: 0x6ba538d9875fb74897b66349fa1f402af9640eaf950b454639ac915e518af9d0
dynssz only         [4237 ms / 3250 ms / 7548 ms]        success         Root: 0x6ba538d9875fb74897b66349fa1f402af9640eaf950b454639ac915e518af9d0
dynssz + fastssz    [ 546 ms /  493 ms / 6698 ms]        success         Root: 0x6ba538d9875fb74897b66349fa1f402af9640eaf950b454639ac915e518af9d0

## minimal preset / BeaconBlock decode + encode + hash (1000 times)
fastssz only        [    ms /     ms /     ms]   failed (unmarshal error: invalid ssz encoding. first variable element offset indexes into fixed value data)
dynssz only         [  33 ms /   54 ms /   35 ms]        success         Root: 0x91c7a6e6ff839bf1e3c0f050d0990a99fbe25c1d7debd9139093d9fce1913068
dynssz + fastssz    [  19 ms /   31 ms /   29 ms]        success         Root: 0x91c7a6e6ff839bf1e3c0f050d0990a99fbe25c1d7debd9139093d9fce1913068

## minimal preset / BeaconState decode + encode + hash (1000 times)
fastssz only        [    ms /     ms /     ms]   failed (unmarshal error: incorrect size)
dynssz only         [ 379 ms /  454 ms /  520 ms]        success         Root: 0x858e5785166165d1ff92bba1b11d619babf57fafc0ef9457afe4cf7469fb93c2
dynssz + fastssz    [  77 ms /   72 ms /  505 ms]        success         Root: 0x858e5785166165d1ff92bba1b11d619babf57fafc0ef9457afe4cf7469fb93c2
```

new typecache:
```
## mainnet preset / BeaconBlock decode + encode + hash (1000 times)
fastssz only        [   1 ms /    0 ms /    8 ms]        success         Root: 0x13515a87c6e7f37b5b39d6d1b69a0819e93b04bb4e6b3c3d57d14d6b7af3b45d
dynssz only         [   2 ms /    2 ms /   10 ms]        success         Root: 0x13515a87c6e7f37b5b39d6d1b69a0819e93b04bb4e6b3c3d57d14d6b7af3b45d
dynssz + fastssz    [   0 ms /    0 ms /    8 ms]        success         Root: 0x13515a87c6e7f37b5b39d6d1b69a0819e93b04bb4e6b3c3d57d14d6b7af3b45d

## mainnet preset / BeaconState decode + encode + hash (1000 times)
fastssz only        [ 506 ms /  452 ms / 7016 ms]        success         Root: 0x6ba538d9875fb74897b66349fa1f402af9640eaf950b454639ac915e518af9d0
dynssz only         [2143 ms / 1190 ms / 7355 ms]        success         Root: 0x6ba538d9875fb74897b66349fa1f402af9640eaf950b454639ac915e518af9d0
dynssz + fastssz    [ 556 ms /  474 ms / 6791 ms]        success         Root: 0x6ba538d9875fb74897b66349fa1f402af9640eaf950b454639ac915e518af9d0

## minimal preset / BeaconBlock decode + encode + hash (1000 times)
fastssz only        [    ms /     ms /     ms]   failed (unmarshal error: invalid ssz encoding. first variable element offset indexes into fixed value data)
dynssz only         [   3 ms /    2 ms /   12 ms]        success         Root: 0x91c7a6e6ff839bf1e3c0f050d0990a99fbe25c1d7debd9139093d9fce1913068
dynssz + fastssz    [   2 ms /    1 ms /   12 ms]        success         Root: 0x91c7a6e6ff839bf1e3c0f050d0990a99fbe25c1d7debd9139093d9fce1913068

## minimal preset / BeaconState decode + encode + hash (1000 times)
fastssz only        [    ms /     ms /     ms]   failed (unmarshal error: incorrect size)
dynssz only         [  65 ms /   35 ms /  265 ms]        success         Root: 0x858e5785166165d1ff92bba1b11d619babf57fafc0ef9457afe4cf7469fb93c2
dynssz + fastssz    [  46 ms /   26 ms /  301 ms]        success         Root: 0x858e5785166165d1ff92bba1b11d619babf57fafc0ef9457afe4cf7469fb93c2
```